% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataEconomicClassification.R
\name{dataEconomicClassification}
\alias{dataEconomicClassification}
\title{Get data about a set of government functions and economic classifications}
\usage{
dataEconomicClassification(year = 2020)
}
\arguments{
\item{year}{number}
}
\value{
tibble.
}
\description{
Get data about a set of government functions and economic classifications
}
\examples{
dataEconomicClassification(year=2020)
}
