% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataEconomicSubFunction.R
\name{dataEconomicSubFunction}
\alias{dataEconomicSubFunction}
\title{Get data about a set of government sub-functions and economic classifications}
\usage{
dataEconomicSubFunction(year = 2020, sel_function = "Educação")
}
\arguments{
\item{year}{number}
}
\value{
tibble.
}
\description{
Get data about a set of government sub-functions and economic classifications
}
\examples{
dataEconomicSubFunction(year=2020)
}
