% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExpenseFlow.R
\name{dataExpenseFlow}
\alias{dataExpenseFlow}
\title{Get data about a set of government functions used as support for sankey graphs}
\usage{
dataExpenseFlow(year = 2020)
}
\arguments{
\item{year}{number}
}
\value{
data for sankey graph.
}
\description{
Get data about a set of government functions used as support for sankey graphs
}
\examples{
dataExpenseFlow(year=2020)
}
