% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataYearlyDistributionPercentage.R
\name{dataYearlyDistributionPercentage}
\alias{dataYearlyDistributionPercentage}
\title{Get time series data for top government functions}
\usage{
dataYearlyDistributionPercentage()
}
\value{
tibble.
}
\description{
Get time series data for top government functions
}
\examples{
dataYearlyDistributionPercentage()
}
