% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineActionCluster.R
\name{defineActionCluster}
\alias{defineActionCluster}
\title{Define clusters for govenrment actions based on total expenses}
\usage{
defineActionCluster(year = 2020)
}
\arguments{
\item{year}{number}
}
\value{
tibble with clusters.
}
\description{
Define clusters for govenrment actions based on total expenses
}
\examples{
defineActionCluster(year=2020)
}
