% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphEconomicClassification.R
\name{graphEconomicClassification}
\alias{graphEconomicClassification}
\title{Generate a ranking graph for economic classification}
\usage{
graphEconomicClassification(.data, free_scale = TRUE)
}
\arguments{
\item{.data}{tibble}

\item{free_scale}{boolean. Indicates if the graph will be presented in free scale}
}
\value{
ranking graph.
}
\description{
Generate a ranking graph for economic classification
}
\examples{
dataEconomicClassification(year=2020) \%>\% graphEconomicClassification()
}
