% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphEconomicSubFunction.R
\name{graphEconomicSubFunction}
\alias{graphEconomicSubFunction}
\title{Generate a ranking graph for economic classification for government sub-function}
\usage{
graphEconomicSubFunction(.data, free_scale = TRUE)
}
\arguments{
\item{.data}{tibble}

\item{free_scale}{boolean. Indicates if the graph will be presented in free scale}
}
\value{
ranking graph.
}
\description{
Generate a ranking graph for economic classification for government sub-function
}
\examples{
dataEconomicSubFunction(year=2020) \%>\% graphEconomicSubFunction()
}
