% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphExpenseFlow.R
\name{graphExpenseFlow}
\alias{graphExpenseFlow}
\title{Generate a sankey graph for a given dataset}
\usage{
graphExpenseFlow(.data)
}
\arguments{
\item{.data}{tibble}
}
\value{
box plot graph.
}
\description{
Generate a sankey graph for a given dataset
}
\examples{
dataExpenseFlow(year=2020) \%>\% graphExpenseFlow()
}
