% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphGovernmentAction.R
\name{graphGovernmentAction}
\alias{graphGovernmentAction}
\title{Generate a distribution graph for a given dataset}
\usage{
graphGovernmentAction(.data, cluster = NULL, action = "Juros")
}
\arguments{
\item{.data}{tibble}

\item{cluster}{number that identifies the clusters to be shown at the graph}

\item{action}{vector that identifies the government action to be highlighted}
}
\value{
jitter graph.
}
\description{
Generate a distribution graph for a given dataset
}
\examples{
defineActionCluster(year=2020) \%>\% graphGovernmentAction()
}
