% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphTimeSeries.R
\name{graphTimeSeries}
\alias{graphTimeSeries}
\title{Generate a time series graph for for government functions}
\usage{
graphTimeSeries(.data)
}
\arguments{
\item{.data}{tibble}
}
\value{
time series graph.
}
\description{
Generate a time series graph for for government functions
}
\examples{
dataTimeSeries(sel_function= c("Saúde", "Educação")) \%>\% graphTimeSeries()
}
