% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphYearlyDistributionPercentage.R
\name{graphYearlyDistributionPercentage}
\alias{graphYearlyDistributionPercentage}
\title{Generate a time series graph for top government functions}
\usage{
graphYearlyDistributionPercentage(.data)
}
\arguments{
\item{.data}{tibble}
}
\value{
time series graph.
}
\description{
Generate a time series graph for top government functions
}
\examples{
dataYearlyDistributionPercentage() \%>\% graphYearlyDistributionPercentage()
}
